python-cli/requirements.txt
pyserial>=3.5
sounddevice>=0.5.3
numpy>=2.0.0
opuslib>=3.0.1
click>=8.1.7
PySSTV>=0.1.0
Pillow>=10.0.0
scipy>=1.11.0
pyaprs==0.0.1